Move helper functions into own module
