+++ .eslintignore
@@ -1,2 +1,2 @@
-build.js
+build
 coverage
