+++ .gitignore
@@ -2,5 +2,5 @@ node_modules
 npm-debug.log
 *.swp
 *.swo
-build.js
+build
 coverage
