+++ .travis.yml
@@ -9,6 +9,7 @@ before_script: greenkeeper-lockfile-update
 script:
   - npm run lint
   - npm test -- --coverage
+  - npm run build
 after_script: greenkeeper-lockfile-upload
 after_success:
   - bash <(curl -s https://codecov.io/bash)
