+++ src/constants.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  REQUIRE: 'require',
+  MODULE: 'module',
+  EXPORTS: 'exports',
+  DEFINE: 'define'
+};
